+++ Backend/src/main/java/edu/cit/hapongo/repository/LessonQuizRepository.java
@@ -4,6 +4,6 @@ import edu.cit.hapongo.model.LessonQuiz;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface LessonQuizRepository extends JpaRepository<LessonQuiz, Integer> {
-    List<LessonQuiz> findByLesson_LessonId(int lessonId);
+public interface LessonQuizRepository extends JpaRepository<LessonQuiz, Long> {
+    List<LessonQuiz> findByLesson_LessonId(long lessonId);
 }
