+++ Backend/src/main/java/edu/cit/hapongo/repository/UserLessonRepository.java
@@ -8,13 +8,13 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserLessonRepository extends JpaRepository<UserLesson, Integer> {
+public interface UserLessonRepository extends JpaRepository<UserLesson, Long> {
 
     // Get a list of UserLesson by userId from the User entity
-    List<UserLesson> findByUser_UserId(int userId);
+    List<UserLesson> findByUser_UserId(long userId);
 
     // Find a specific UserLesson by userId and lessonId
-    Optional<UserLesson> findByUser_UserIdAndLesson_LessonId(int userId, int lessonId);
+    Optional<UserLesson> findByUser_UserIdAndLesson_LessonId(long userId, long lessonId);
 }
