+++ Backend/src/main/java/edu/cit/hapongo/repository/UserRepository.java
@@ -4,8 +4,8 @@ import edu.cit.hapongo.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
     Optional<User> findByName(String name);
-    Optional<User> findById(Integer id);
+    Optional<User> findById(Long id);
 }
